ml-training/requirements.txt
# TensorFlow and ML Dependencies
tensorflow>=2.15.0
tensorflowjs>=4.15.0
keras>=2.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# Data Augmentation
albumentations>=1.3.0
imgaug>=0.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
python-dotenv>=1.0.0

# Model Utilities
h5py>=3.9.0
tensorboard>=2.15.0

# Dataset Management
kaggle>=1.5.16
gdown>=4.7.1
